ca_automation/requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
tensorflow-cpu==2.16.1
transformers==4.44.2
torch==2.3.1
pytesseract==0.3.10
Pillow==10.4.0
plotly==5.24.1
psycopg2-binary==2.9.9
web3==6.20.0
requests==2.32.3
markdown==3.7
gym==0.26.2
opencv-python==4.10.0
pyyaml==6.0.2